% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_assignments_genes.R
\name{cluster_assignments_genes}
\alias{cluster_assignments_genes}
\title{cluster_assignments_genes}
\usage{
cluster_assignments_genes(cluster_genes, Prot_annotation = NULL)
}
\arguments{
\item{cluster_genes}{A data frame containing cluster assignments for genes.}

\item{Prot_annotation}{A data frame with the annotation names of the proteins. Should have a column called Feature_ID.}
}
\value{
A data frame containing cluster assignments and HMDB/Symbol.
}
\description{
Cluster assignments for proteins and genes.
}
