% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_assignments_genes.R
\name{cluster_assignments_genes}
\alias{cluster_assignments_genes}
\title{cluster_assignments_genes}
\usage{
cluster_assignments_genes(cluster_genes, Prot_annotation = NULL)
}
\arguments{
\item{hclustTree}{(hierarchical clustering object)}

\item{Genes_Annotation}{(data frame with de symbol genes of the Proteins or RNAseq: Should have a column called Feature_ID}
}
\value{
result_list data frame containing cluster_assignments and HMDB/Symbol
}
\description{
cluster Assignments for proteins and genes
}
