% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_assignments_metabolites.R
\name{cluster_assignments_metabolites}
\alias{cluster_assignments_metabolites}
\title{cluster_assignments_metabolites}
\usage{
cluster_assignments_metabolites(cluster_metabolites, metab_annotation = NULL)
}
\arguments{
\item{cluster_metabolites}{A data frame containing cluster assignments for metabolites}

\item{metab_annotation}{A data frame with the annotation names of the metabolites. Should have a column called Feature_ID.}
}
\value{
A data frame containing cluster assignments and HMDB/Symbol.
}
\description{
cluster Assignments for metabolites
}
