% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Eigengenes.R
\name{Eigengenes}
\alias{Eigengenes}
\title{Eigengenes}
\usage{
Eigengenes(
  Expression_mat = Expression_mat,
  cluster_assignments = cluster_assignments
)
}
\value{
A feature matrix with the first principal component of each metabolomic/protein cluster
}
\description{
moduleEigengenes function by WGCNA
}
