% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_FeaturesAnnot_correlation.R
\name{FeaturesAnnot_correlation}
\alias{FeaturesAnnot_correlation}
\title{FeaturesAnnot_correlation}
\usage{
FeaturesAnnot_correlation(
  Cor_Datai_Dataj,
  cluster_assignments_D1,
  cluster_assignments_D2,
  load_data1 = load_data1,
  load_data2 = load_data2,
  top_n = 5
)
}
\arguments{
\item{Cor_Datai_Dataj}{A data frame with the first principal component of each protein cluster and their correlations.}

\item{cluster_assignments_D1}{A data frame containing cluster assignments for metabolites.}

\item{cluster_assignments_D2}{A data frame containing cluster assignments and Enrichr terms for proteins.}

\item{top_n}{The number of top correlations to select.}

\item{ExpressionD2}{A feature matrix (e.g., protein expression) with samples in rows and features (e.g., proteins) in columns. Row names must be unique.}

\item{ExpressionD1}{A feature matrix (e.g., metabolite expression) with samples in rows and features (e.g., metabolites) in columns. Row names must be unique.}
}
\value{
A list containing important features and correlation matrices.
}
\description{
Calculate correlation between the features of top correlated modules.
}
