% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(Expression_mat = Expression_mat)
}
\arguments{
\item{Expression_mat}{A feature matrix (e.g., gene expression) with samples in rows and features (e.g., genes) in columns. Row names must be unique.}
}
\value{
A prepossessed data matrix.
}
\description{
A function to load and preprocess the expression data.
}
