% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation_matrices_list.R
\name{annotation_matrices_list}
\alias{annotation_matrices_list}
\title{annotation_matrices_list}
\usage{
annotation_matrices_list(
  Cor_Prot_Metab,
  cluster_assignments_Prot,
  cluster_assignments_metab,
  Prot_annotation = NULL,
  metab_annotation = NULL,
  top_n = 5
)
}
\arguments{
\item{Cor_Prot_Metab}{A data frame with the first principal component of each protein cluster and their correlations.}

\item{cluster_assignments_Prot}{A data frame containing cluster assignments and Enrichr terms for proteins.}

\item{cluster_assignments_metab}{A data frame containing cluster assignments for metabolites.}

\item{Prot_annotation}{A data frame with the annotation names of the proteins. Should have a column called Symbol.}

\item{metab_annotation}{A data frame with the annotation names of the metabolites. Should have a column called HMDB.}

\item{top_n}{The number of top correlations to select. Default is 5.}
}
\value{
A list containing annotation matrices for the top correlated modules.
}
\description{
Calculate correlation between the features of top correlated modules.
}
