% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Assigment_genes_enrichr.R
\name{Assigment_genes_enrichr}
\alias{Assigment_genes_enrichr}
\title{Assigment_genes_enrichr}
\usage{
Assigment_genes_enrichr(
  cluster_assignments_ProtGenes,
  database = "GO_Biological_Process_2023"
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
