% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_assignments.R
\name{cluster_assignments}
\alias{cluster_assignments}
\title{cluster_assignments}
\usage{
cluster_assignments(cluster, PhenoData = NULL, selected_columns = NULL)
}
\arguments{
\item{cluster}{A data frame containing cluster assignments.}

\item{PhenoData}{A data frame with the annotation names of the Dataset. Should have a column called Feature_ID.}
}
\value{
A data frame containing cluster assignments and selected columns by user.
}
\description{
Cluster assignments with annotation.
}
