% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_assignments_genes.R
\name{cluster_assignments_genes}
\alias{cluster_assignments_genes}
\title{cluster_assignments_genes}
\usage{
cluster_assignments_genes(cluster_genes, Prot_annotation = NULL)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
