% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_assignments_metabolites.R
\name{cluster_assignments_metabolites}
\alias{cluster_assignments_metabolites}
\title{cluster_assignments_metabolites}
\usage{
cluster_assignments_metabolites(cluster_metabolites, metab_annotation = NULL)
}
\arguments{
\item{hclustTree}{(hierarchical clustering object)}

\item{Metabolites_annotation}{(data frame with de annotation names of the metabolites: Should have a column called HMDB)}
}
\value{
result_list data frame containing cluster_assignments and HMDB/Symbol
}
\description{
cluster Assignments for metabolites
}
