% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Eigengenes.R
\name{Eigengenes}
\alias{Eigengenes}
\title{Eigengenes}
\usage{
Eigengenes(
  feature_mat_t = feature_mat_t,
  cluster_assignments = cluster_assignments
)
}
\value{
The return value matrix.
}
\description{
A Eigengenes function
}
