% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Eigengenes.R
\name{Eigengenes}
\alias{Eigengenes}
\title{Eigengenes}
\usage{
Eigengenes(load_data = load_data, cluster_assignments = cluster_assignments)
}
\arguments{
\item{cluster_assignments}{A vector of cluster assignments for each feature.}

\item{Expression_mat}{A feature matrix (e.g., gene expression) with samples in rows and features (e.g., genes) in columns. Row names must be unique.}
}
\value{
A list containing:
\item{module_eigenmetab_List_Me}{The full result from WGCNA::moduleEigengenes.}
\item{module_eigenmetab_Me}{The matrix of module eigengenes.}
\item{feature_mat_t}{The filtered and scaled feature matrix.}
}
\description{
Calculates module eigengenes using the WGCNA package.
}
