% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_FeaturesAnnot_correlation.R
\name{FeaturesAnnot_correlation}
\alias{FeaturesAnnot_correlation}
\title{FeaturesAnnot_correlation}
\usage{
FeaturesAnnot_correlation(
  Cor_Prot_Metab,
  cluster_assignments_Prot,
  cluster_assignments_metab,
  ExpressionProt_mat = ExpressionProt_mat,
  ExpressionMetab_mat = ExpressionMetab_mat,
  top_n = 5
)
}
\arguments{
\item{Cor_Prot_Metab}{A data frame with the first principal component of each protein cluster and their correlations.}

\item{cluster_assignments_Prot}{A data frame containing cluster assignments and Enrichr terms for proteins.}

\item{cluster_assignments_metab}{A data frame containing cluster assignments for metabolites.}

\item{ExpressionProt_mat}{A feature matrix (e.g., protein expression) with samples in rows and features (e.g., proteins) in columns. Row names must be unique.}

\item{ExpressionMetab_mat}{A feature matrix (e.g., metabolite expression) with samples in rows and features (e.g., metabolites) in columns. Row names must be unique.}

\item{top_n}{The number of top correlations to select.}
}
\value{
A list containing important features and correlation matrices.
}
\description{
Calculate correlation between the features of top correlated modules.
}
