% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_FeaturesAnnot_correlation.R
\name{FeaturesAnnot_correlation}
\alias{FeaturesAnnot_correlation}
\title{FeaturesAnnot_correlation}
\usage{
FeaturesAnnot_correlation(
  Cor_Datai_Dataj,
  cluster_assignments_D1,
  cluster_assignments_D2,
  load_data1 = load_data1,
  load_data2 = load_data2,
  top_n = 5
)
}
\arguments{
\item{Cor_Datai_Dataj}{A data frame with the first principal component of each protein cluster and their correlations.}

\item{cluster_assignments_D1}{A data frame containing cluster assignments for metabolites.}

\item{cluster_assignments_D2}{A data frame containing cluster assignments and Enrichr terms for proteins.}

\item{load_data1}{A prepossessed data matrix resulting from "load data" function from Data1.}

\item{load_data2}{A prepossessed data matrix resulting from "load data" function from Data2.}

\item{top_n}{The number of top correlations to select.}
}
\value{
A list containing important features and correlation matrices.
}
\description{
Calculate correlation between the features of top correlated modules.
}
