% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_FeaturesAnnot_correlation.R
\name{FeaturesAnnot_correlation}
\alias{FeaturesAnnot_correlation}
\title{FeaturesAnnot_correlation}
\usage{
FeaturesAnnot_correlation(
  Cor_Prot_Metab,
  cluster_assignments_Prot,
  cluster_assignments_metab,
  Prot_t,
  metab_t,
  top_n = 5
)
}
\arguments{
\item{Cor_Prot_Metab}{A feature matrix with the first principal component of each protein cluster}

\item{cluster_assignments_Prot}{data frame containing cluster_assignments and Enrichr terms}

\item{cluster_assignments_metab}{data frame containing cluster_assignments}

\item{Prot_t}{A feature matrix (e.g. gene expression) with samples in columns and features (e.g. genes) in rows Row names must be unique.}

\item{metab_t}{A feature matrix (e.g. gene expression) with samples in columns and features (e.g. metabolites) in rows Row names must be unique.}

\item{Metabolites_annotation}{(data frame with de annotation names of the metabolites: Should have a column called KEGG)}
}
\value{
Important features
}
\description{
Calculate correlation between the features of top correlated modules
}
