% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hierarchical_cluster.R
\name{hierarchical_cluster}
\alias{hierarchical_cluster}
\title{hierarchical_cluster}
\usage{
hierarchical_cluster(parcor_mat, tom = TRUE, min_module_size = 10)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
