% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hierarchical_cluster.R
\name{hierarchical_cluster}
\alias{hierarchical_cluster}
\title{hierarchical_cluster}
\usage{
hierarchical_cluster(parcor_mat, tom = TRUE, min_module_size = 10)
}
\arguments{
\item{parcor_mat}{A partial correlation matrix from glassoFast$wi: Estimated inverse covariance matrix}

\item{tom}{TRUE: topological overlap measure; FALSE: partial correlation matrix; default is TRUE.}

\item{min_module_size}{Smallest modules to be generated; default is 10}
}
\value{
result_list List containing TOM_diss (TOM dissimilarity; NA if argument TOM = FALSE), hclustTree (hierarchical clustering object), dynamicMods (index of module membership for features), and mod_count (number of modules).
}
\description{
Perform hierarchical clustering of a partial correlation matrix. By default uses the topological overlap measure (TOM)
}
