% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(file_name)
}
\arguments{
\item{file_name}{The name of the fle to upload.}
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
