% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Modules_correlation.R
\name{Modules_correlation}
\alias{Modules_correlation}
\title{Modules_correlation}
\usage{
Modules_correlation(eigengenes_list, cluster_list, threshold = 0.5)
}
\arguments{
\item{threshold}{A numeric value to filter correlations. Default is 0.5}

\item{eigengenes}{A feature matrix with the first principal component of each cluster}

\item{cluster_assignments}{data frame containing cluster_assignments}
}
\value{
Return the Top_correlations between Prot and metab eigengenes, graph of red, correlation plot
}
\description{
Calculates correlations between omic modules.
}
