% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_partial_cors.R
\name{partial_cors}
\alias{partial_cors}
\title{partial_cors}
\usage{
partial_cors(Expression_mat)
}
\arguments{
\item{feature_mat_t}{A feature matrix (e.g. gene expression) with samples in rows and features (e.g. genes) in columns. Row names must be unique.}
}
\value{
partial_cor_mat A partial correlation matrix with NA's in the diagonal.
}
\description{
Calculates partial correlations based on graphical lasso (https://www.rdocumentation.org/packages/glassoFast/versions/1.0/topics/glassoFast).
}
