% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_partial_cors.R
\name{partial_cors}
\alias{partial_cors}
\title{partial_cors}
\usage{
partial_cors(load_data, rho = 0.25)
}
\arguments{
\item{load_data}{A prepossessed data matrix resulting from "load data" function.}

\item{rho}{The regularization parameter for lasso. (a non-negative value or a p by p matrix of regularization parameters).}
}
\value{
partial_cor_mat A partial correlation matrix with NA's in the diagonal.
}
\description{
Calculates partial correlations based on graphical lasso (https://www.rdocumentation.org/packages/glassoFast/versions/1.0/topics/glassoFast).
}
