% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_perform_classification.R
\name{perform_classification}
\alias{perform_classification}
\title{perform_classification}
\usage{
perform_classification(
  eigengene_data,
  metadata,
  phenotype_variable,
  significance_threshold = 0.05
)
}
\arguments{
\item{eigengene_data}{A data frame of eigengenes organized with patient IDs in rows and variables in columns.}

\item{metadata}{A data frame containing metadata, with a column call "Sample" that matches patient IDs in eigengene_data.}

\item{phenotype_variable}{The variable selected by the user in the Shiny app (response variable).}

\item{significance_threshold}{A numeric value to filter p-value. Default is 00.5.}
}
\value{
A data frame with metrics such as AUC, Accuracy, and Error Rate for each binary classification.
}
\description{
Performs classification using different methods such as Welch’s T-test, Random Forest, and K-Nearest Neighbors.
}
